Add tables back for STI models
Added badges, badge_sets, earned_badges, teams, and team_memberships.

Where we left off, we weren't really doing anything but storing two
separate models in the same table with a type column to segment them.
There wasn't a huge benefit to that, and as we're doing more in the
database I think we would end up regretting the decision.

We did end up with some successful changes due to this process: for
example, submissions and tasks are now polymorphic so they'll work for
assignments, badges, and challenges.