Fix routing, move query into User model.
Fixed the routing. The query to obtain the chart data will now return
some data based on the user id provided. However, I don't believe the
current query is correct because it does not take the current course
into account. I'll add that in a subsequent commit.

However, now would be a good time for someone to generate a graph using
the returned JSON.

See #197.
See #145.