Correct assigment scores query, all at once.
Select the assignment scores all at once. The top ten and bottom ten are
determined by the data attributes attached to the table cell that will
house the bar chart. We download a summary for all the students.
Transform it into raw material for the bar chart for each student and
render the charts when the tables are toggled.

Closes #262.