Embed predictor JSON data into HTML.
Embed the predictor json data into the student dashboard HTML. This
eliminates the stutter when the browser returns to the server for the
predictor JSON when the student dashboard page loads. The predictor
chart is at the top of the student page, so the student is looking at a
blank page until it is rendered. They shouldn't have to wait.

Also, removed duplicate predictor `get` action from user controller.

See #431.