Use join for badges on student dashboard.
Use an ActiveRecord join to select the badges for the student dashboard.
This makes a single trip to to the database to obtain all of the records
necessary to display the badges table of the student dashboard. No
longer using the momoization within badges, so it may not be necessary
if it is not used elsewhere.

Closes #477.