Student dashboard views.
Generate the student dashboard using PostgreSQL views. Primarily a view
called membership_calculation that gathers all the summaries that are
specific to the given student and course of the current dashboard.

These summaries are then used in the `current_student_data` where
indivdual queries for each summary would have been run.

Also created `membership_scores` that contains scores grouped by
assignment type, a query used by the predictor. The generation of the
JSON needed for the predictor has been moved from `User` to
`StudentData`.

Also created a `released_grades` view to simplify some of the other
queries. It filters grades using the released or graded and no release
necessary condition that is found throughout the code. I'm not using
that view enough in the SQL, there are more opportunties to use it.